Chnaged to longs so health shows up

+++ client/src/main/java/edu/unh/cs/cs619/bulletzone/ui/ViewFactory.java
@@ -52,7 +52,7 @@ public class ViewFactory {
         layers[3] = context.getDrawable(R.drawable.empty);
 
         int dir = value % 10;
-        int health = (value/10) % 1000;
+        long health = (value/10) % 1000;
         int tankID = (value/10000) % 1000;
         int objectType = (value/ 10000000) % 10;
         int terrainType = value / 100000000;
@@ -142,7 +142,7 @@ public class ViewFactory {
         //set health
         if(health != 0) {
             int hea = R.drawable.empty;
-            int total = -1;
+            long total = -1;
 
             switch (objectType){
                 case TANK:
@@ -161,7 +161,7 @@ public class ViewFactory {
 
             }
 
-            long percentage = (long)health/total;
+            long percentage = health/total;
             if(percentage == 0.0)
                 hea = R.drawable.health_0;
             else if(percentage > 0.0 && percentage <= 0.3)
